requirements.txt
# Core MCP framework and dependencies
fastmcp>=2.12.0,<3.0.0
pydantic>=2.0.0,<3.0.0

# HTTP and networking
aiohttp>=3.8.0,<4.0.0
httpx>=0.24.0,<1.0.0
websockets>=11.0.0,<12.0.0

# Web framework (for potential future REST API)
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.21.0,<1.0.0

# Configuration and utilities
python-dotenv>=1.0.0,<2.0.0
tomli>=2.0.0,<3.0.0
python-dateutil>=2.8.2,<3.0.0
anyio>=4.5.0,<5.0.0

# Logging and debugging
structlog>=23.1.0,<24.0.0

# Type hints support
typing-extensions>=4.5.0,<5.0.0
